pyproject.toml
[tool.poetry]
name = "gsg"
version = "0.1.0"
description = "GSG Project"
authors = ["Teodor Angelov <teodorangel@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gsg" },
  { include = "crawler" },
  { include = "processors" }
]

[tool.poetry.dependencies]
fire = "^0.5"
python = ">=3.9,<4.0"
scrapy = "^2.12.0"
beautifulsoup4 = "^4.13.4"
langchain = "^0.3.24"
langchain-community = "*"
pydantic-ai = "^0.1.4"
psycopg = {extras = ["binary"], version = "^3.2.6"}
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
tiktoken = "*"
python-dotenv = "^1.1.0"
rich = "^14.0.0"
pinecone = "^2.2"
requests = "^2.32.3"
langchain-openai = "^0.3.14"
fastapi = "^0.101"
uvicorn = "^0.23"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.1.0,<8.2.0"
pytest-asyncio = ">=0.23.0,<0.24.0"
pytest-cov = ">=4.1.0,<4.2.0"
pytest-httpx = "^0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
# Install development dependencies via pip install .[dev]
dev = [
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "requests",
  "fastapi",
  "langchain-openai",
  "fire"
]
